Move Notifications to Left Nav, Reorder and add divider
