+++ assets/tests/components/nav/leftNav.test.tsx
@@ -9,6 +9,7 @@ import {
   initialState,
   openLateView,
   openSwingsView,
+  openNotificationDrawer,
   OpenView,
 } from "../../../src/state"
 import { openDrift } from "../../../src/helpers/drift"
@@ -178,5 +179,57 @@ describe("LeftNav", () => {
     await user.click(result.getByTitle("About Skate"))
 
     expect(displayHelp).toHaveBeenCalled()
+  })
+
+  test("clicking notifications icon toggles notifications drawer", async () => {
+    const dispatch = jest.fn()
+    const user = userEvent.setup()
+    const result = render(
+      <StateDispatchProvider state={initialState} dispatch={dispatch}>
+        <BrowserRouter>
+          <LeftNav defaultToCollapsed={false} dispatcherFlag={false} />
+        </BrowserRouter>
+      </StateDispatchProvider>
+    )
+
+    await user.click(result.getByTitle("Notifications"))
+
+    expect(dispatch).toHaveBeenCalledWith(openNotificationDrawer())
+  })
+
+  test("notifications icon gets active class when notifications drawer is open", () => {
+    const dispatch = jest.fn()
+    const result = render(
+      <StateDispatchProvider
+        state={{ ...initialState, notificationDrawerIsOpen: true }}
+        dispatch={dispatch}
+      >
+        <BrowserRouter>
+          <LeftNav defaultToCollapsed={false} dispatcherFlag={false} />
+        </BrowserRouter>
+      </StateDispatchProvider>
+    )
+
+    expect(result.getByTitle("Notifications").children[0]).toHaveClass(
+      "m-top-nav__notifications-icon--active"
+    )
+  })
+
+  test("notifications icon doesn't get active class when notifications drawer is close", () => {
+    const dispatch = jest.fn()
+    const result = render(
+      <StateDispatchProvider
+        state={{ ...initialState, notificationDrawerIsOpen: false }}
+        dispatch={dispatch}
+      >
+        <BrowserRouter>
+          <LeftNav defaultToCollapsed={false} dispatcherFlag={false} />
+        </BrowserRouter>
+      </StateDispatchProvider>
+    )
+
+    expect(result.getByTitle("Notifications").children[0]).not.toHaveClass(
+      "m-top-nav__notifications-icon--active"
+    )
   })
 })
