+++ assets/tests/components/nav/topNav.test.tsx
@@ -3,67 +3,12 @@ import { render } from "@testing-library/react"
 import TopNav from "../../../src/components/nav/topNav"
 import userEvent from "@testing-library/user-event"
 import { StateDispatchProvider } from "../../../src/contexts/stateDispatchContext"
-import {
-  initialState,
-  openNotificationDrawer,
-  OpenView,
-} from "../../../src/state"
+import { initialState } from "../../../src/state"
 import { BrowserRouter } from "react-router-dom"
 import "@testing-library/jest-dom"
 import * as browser from "../../../src/models/browser"
 
 describe("TopNav", () => {
-  test("clicking notifications icon toggles notifications drawer", async () => {
-    const dispatch = jest.fn()
-    const user = userEvent.setup()
-    const result = render(
-      <StateDispatchProvider state={initialState} dispatch={dispatch}>
-        <BrowserRouter>
-          <TopNav />
-        </BrowserRouter>
-      </StateDispatchProvider>
-    )
-
-    await user.click(result.getByTitle("Notifications"))
-
-    expect(dispatch).toHaveBeenCalledWith(openNotificationDrawer())
-  })
-
-  test("notifications icon gets active class when notifications drawer is open", () => {
-    const dispatch = jest.fn()
-    const result = render(
-      <StateDispatchProvider
-        state={{ ...initialState, openView: OpenView.NotificationDrawer }}
-        dispatch={dispatch}
-      >
-        <BrowserRouter>
-          <TopNav />
-        </BrowserRouter>
-      </StateDispatchProvider>
-    )
-
-    expect(result.getByTitle("Notifications").children[0]).toHaveClass(
-      "m-top-nav__notifications-icon--active"
-    )
-  })
-
-  test("notifications icon doesn't get active class when notifications drawer is closed", () => {
-    const dispatch = jest.fn()
-    const result = render(
-      <StateDispatchProvider
-        state={{ ...initialState, openView: OpenView.None }}
-        dispatch={dispatch}
-      >
-        <BrowserRouter>
-          <TopNav />
-        </BrowserRouter>
-      </StateDispatchProvider>
-    )
-
-    expect(result.getByTitle("Notifications").children[0]).not.toHaveClass(
-      "m-top-nav__notifications-icon--active"
-    )
-  })
 
   test("refresh button reloads the page", async () => {
     const reloadSpy = jest
